+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Version 2.4.8
 * Removed direct GO mock replacement
+* Added mock support for custom vegetation
 
 ## Version 2.4.7
 * Fixed mock resolving of certain components (piece place effects for example)